+++ lib/mirage_configure_libvirt.mli
@@ -1,0 +1,6 @@
+open Rresult
+
+val filename : name:string -> Fpath.t
+val configure_main : root:string -> name:string -> (unit, [> R.msg ]) result
+val configure_virtio : root:string -> name:string -> (unit, [> R.msg ]) result
+val clean : name:string -> (unit, [> R.msg ]) result