+++ test/functoria-e2e/app/app.ml
@@ -1,0 +1,2 @@
+let start =
+  Fmt.pr "Success: vote=%s hello=%s\n%!" Key_gen.(vote ()) Key_gen.(hello ())